Split out DslSpec base class to a separate file
And add some comments to details what is going on here
